refactor: extract fat controllers into thin adapters + dedicated services
Adopts the emhub stability pattern: controllers are thin HTTP adapters,
services own all business logic and data access. No Prisma imports in controllers.

**Phase 1 — Fat controller extraction (4 controllers → 4 new services):**

- pipeline.controller.ts (264→61 lines) → pipeline.service.ts
  Extracted: 10 Prisma calls, success rate calc, config transform, audit logging
  Also moved inline Prisma query from pipeline.routes.ts into service

- dashboard.controller.ts (466→31 lines) → dashboard.service.ts
  Extracted: 5 Prisma queries, cache stats, metrics aggregation, provider comparison

- auth.controller.ts (245→47 lines) → user.service.ts
  Extracted: 7 Prisma calls, bcrypt hashing, token generation, audit logging

- notification.controller.ts (244→54 lines) → notification-preference.service.ts
  Extracted: 6 Prisma calls, channel config, broadcast logic

**Phase 2 — AI route split (778→26 lines in index):**

- routes/ai/index.ts split into 3 focused sub-modules:
  - config.ts: health, personas, autonomy preferences, provider config
  - analysis.ts: RCA, categorization, summarization, enrichment, costs/stats
  - chat.ts: SSE chat, session CRUD, action confirmation

**Phase 3 — Convention update:**

- SENIOR_ENGINEER.md updated with "Thin Controller Rule" convention
- Before-merge checklist now requires "no prisma imports in controllers"

All tests pass (60 backend + 115 frontend), typecheck clean, lint clean, schema parity OK.
