+++ backend/src/controllers/dashboard.controller.ts
@@ -1,131 +1,20 @@
 /**
- * Dashboard Controller
+ * DashboardController — Thin HTTP adapter.
  *
- * Provides comprehensive dashboard metrics for AI-powered failure analysis
+ * Delegates all metrics aggregation and data access to DashboardService.
+ * No Prisma imports allowed here.
  */
 
 import { Request, Response } from 'express';
-import { prisma } from '../lib/prisma';
-import { getCache } from '../services/ai/cache';
-import { FailureCategory } from '../services/ai/types';
+import { dashboardService } from '../services/dashboard.service';
 
-export interface DashboardMetrics {
-  // Summary stats
-  totalTestsAnalyzed: number;
-  failuresAutoCategorized: number;
-  timeSavedHours: number;
-  aiCostUSD: number;
-  cacheHitRate: number;
-  cacheHits: number;
-  cacheSavingsPercent: number;
-
-  // Metadata
-  lastUpdated: string;
-  timeRange: string;
-
-  // Failure categories breakdown
-  failureCategories: {
-    category: FailureCategory;
-    count: number;
-    percentage: number;
-    color: string;
-  }[];
-
-  // Recent AI failure analysis
-  recentFailures: {
-    id: string;
-    testName: string;
-    errorMessage: string;
-    rootCause: string | null;
-    category: FailureCategory;
-    confidence: number;
-    similarCount: number;
-    filePath: string | null;
-    timestamp: string;
-  }[];
-
-  // AI performance metrics
-  aiPerformance: {
-    avgAnalysisTimeSeconds: number;
-    categorizationAccuracy: number;
-    similarFailuresFound: number;
-    cacheHitRate: number;
-    monthlyBudgetUsed: number;
-    monthlyBudgetTotal: number;
-  };
-
-  // AI provider comparison
-  providers: {
-    name: string;
-    costPer1M: number;
-    contextWindow: string;
-    speed: string;
-    isActive: boolean;
-  }[];
-}
+export type { DashboardMetrics } from '../services/dashboard.service';
 
 export class DashboardController {
-  /**
-   * GET /api/v1/dashboard
-   * Get comprehensive AI dashboard metrics
-   */
   static async getDashboardMetrics(req: Request, res: Response): Promise<void> {
     try {
-      const timeRange = req.query.timeRange || '30d';
-      const startDate = DashboardController.getStartDate(timeRange as string);
-
-      // Fetch all metrics in parallel for performance
-      const [
-        totalTests,
-        failureStats,
-        categorizedFailures,
-        recentFailures,
-        cacheStats,
-        aiCostData,
-      ] = await Promise.all([
-        DashboardController.getTotalTestsCount(startDate),
-        DashboardController.getFailureStats(startDate),
-        DashboardController.getCategorizedFailuresCount(startDate),
-        DashboardController.getRecentFailures(),
-        DashboardController.getCacheStats(),
-        DashboardController.getAICostData(startDate),
-      ]);
-
-      // Calculate derived metrics
-      const totalFailures = failureStats.total;
-      const categorizationRate = totalFailures > 0
-        ? (categorizedFailures / totalFailures) * 100
-        : 100;
-
-      // Calculate time saved (assume 15 min per manual analysis)
-      const timeSavedHours = (categorizedFailures * 15) / 60;
-
-      // Get failure categories breakdown
-      const failureCategories = await DashboardController.getFailureCategoriesBreakdown(startDate);
-
-      // Build dashboard metrics response
-      const metrics: DashboardMetrics = {
-        totalTestsAnalyzed: totalTests,
-        failuresAutoCategorized: categorizedFailures,
-        timeSavedHours: Math.round(timeSavedHours * 10) / 10,
-        aiCostUSD: aiCostData.totalCost,
-        cacheHitRate: cacheStats.hitRate,
-        cacheHits: cacheStats.hits,
-        cacheSavingsPercent: Math.round(cacheStats.savingsPercent),
-        lastUpdated: new Date().toISOString(),
-        timeRange: DashboardController.getTimeRangeLabel(timeRange as string),
-        failureCategories,
-        recentFailures,
-        aiPerformance: {
-          avgAnalysisTimeSeconds: 2.3, // Would be tracked in real implementation
-          categorizationAccuracy: Math.min(96.8, categorizationRate),
-          similarFailuresFound: await DashboardController.getSimilarFailuresCount(),
-          cacheHitRate: cacheStats.hitRate,
-          monthlyBudgetUsed: aiCostData.monthlySpent,
-          monthlyBudgetTotal: aiCostData.monthlyBudget,
-        },
-        providers: DashboardController.getProviderComparison(),
-      };
+      const timeRange = (req.query.timeRange as string) || '30d';
+      const metrics = await dashboardService.getMetrics(timeRange);
 
       res.json({
         success: true,
@@ -137,330 +26,6 @@ export class DashboardController {
         success: false,
         error: error instanceof Error ? error.message : 'Failed to fetch dashboard metrics',
       });
-    }
-  }
-
-  /**
-   * Helper: Get total tests count
-   */
-  private static async getTotalTestsCount(startDate: Date): Promise<number> {
-    try {
-      // This would query your actual test runs table
-      // For now, return a realistic number
-      const count = await prisma.failureArchive.count({
-        where: {
-          lastOccurrence: {
-            gte: startDate,
-          },
-        },
-      });
-
-      // Estimate total tests (assuming ~5% failure rate)
-      return Math.round(count * 20);
-    } catch (error) {
-      console.error('Failed to get total tests count:', error);
-      return 0;
-    }
-  }
-
-  /**
-   * Helper: Get failure statistics
-   */
-  private static async getFailureStats(startDate: Date): Promise<{ total: number }> {
-    try {
-      const total = await prisma.failureArchive.count({
-        where: {
-          lastOccurrence: {
-            gte: startDate,
-          },
-        },
-      });
-
-      return { total };
-    } catch (error) {
-      console.error('Failed to get failure stats:', error);
-      return { total: 0 };
-    }
-  }
-
-  /**
-   * Helper: Get categorized failures count
-   */
-  private static async getCategorizedFailuresCount(startDate: Date): Promise<number> {
-    try {
-      return await prisma.failureArchive.count({
-        where: {
-          lastOccurrence: {
-            gte: startDate,
-          },
-          category: {
-            not: null,
-          },
-        },
-      });
-    } catch (error) {
-      console.error('Failed to get categorized failures count:', error);
-      return 0;
-    }
-  }
-
-  /**
-   * Helper: Get recent failures with AI analysis
-   */
-  private static async getRecentFailures(): Promise<DashboardMetrics['recentFailures']> {
-    try {
-      const failures = await prisma.failureArchive.findMany({
-        where: {
-          category: {
-            not: null,
-          },
-        },
-        orderBy: {
-          lastOccurrence: 'desc',
-        },
-        take: 10,
-        select: {
-          id: true,
-          testName: true,
-          errorMessage: true,
-          rootCause: true,
-          category: true,
-          severity: true,
-          stackTrace: true,
-          lastOccurrence: true,
-        },
-      });
-
-      return failures.map(failure => {
-        // Extract file path from stack trace
-        const filePathMatch = failure.stackTrace?.match(/at\s+(?:.*?\s+)?\(?(\/[^:)]+):\d+:\d+/);
-        const filePath = filePathMatch ? filePathMatch[1] : null;
-
-        // Map severity to category
-        const categoryMap: Record<string, FailureCategory> = {
-          'CRITICAL': 'bug_critical',
-          'HIGH': 'bug_critical',
-          'MEDIUM': 'bug_minor',
-          'LOW': 'bug_minor',
-        };
-        const category = failure.severity ? categoryMap[failure.severity] || 'unknown' : 'unknown';
-
-        return {
-          id: failure.id,
-          testName: failure.testName,
-          errorMessage: failure.errorMessage.substring(0, 100),
-          rootCause: failure.rootCause,
-          category,
-          confidence: 0.85, // Default confidence since we don't have AI scoring yet
-          similarCount: 0, // Would query vector DB for similar failures
-          filePath,
-          timestamp: failure.lastOccurrence.toISOString(),
-        };
-      });
-    } catch (error) {
-      console.error('Failed to get recent failures:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Helper: Get cache statistics
-   */
-  private static async getCacheStats(): Promise<{
-    hits: number;
-    hitRate: number;
-    savingsPercent: number;
-  }> {
-    try {
-      const cache = getCache();
-      const stats = cache.getStats();
-
-      // Calculate savings based on cache hits
-      const savingsPercent = stats.hitRate > 0 ? Math.round(stats.hitRate * 100 * 0.9) : 0;
-
-      return {
-        hits: stats.hits,
-        hitRate: Math.round(stats.hitRate * 100) / 100,
-        savingsPercent,
-      };
-    } catch (error) {
-      console.error('Failed to get cache stats:', error);
-      return { hits: 0, hitRate: 0, savingsPercent: 0 };
-    }
-  }
-
-  /**
-   * Helper: Get AI cost data
-   */
-  private static async getAICostData(_startDate: Date): Promise<{
-    totalCost: number;
-    monthlySpent: number;
-    monthlyBudget: number;
-  }> {
-    try {
-      // This would query the ai_usage table from cost-tracker
-      // For now, return realistic estimates
-      // Mock data - would be real queries in production
-      return {
-        totalCost: 2.34,
-        monthlySpent: 47.20,
-        monthlyBudget: 100,
-      };
-    } catch (error) {
-      console.error('Failed to get AI cost data:', error);
-      return { totalCost: 0, monthlySpent: 0, monthlyBudget: 100 };
-    }
-  }
-
-  /**
-   * Helper: Get failure categories breakdown
-   */
-  private static async getFailureCategoriesBreakdown(
-    startDate: Date
-  ): Promise<DashboardMetrics['failureCategories']> {
-    try {
-      // Fetch all failures to parse category from testName
-      const failures = await prisma.failureArchive.findMany({
-        where: {
-          lastOccurrence: {
-            gte: startDate,
-          },
-        },
-        select: {
-          testName: true,
-        },
-      });
-
-      // Parse category from testName format: TestName_category_index
-      const categoryCounts: Record<string, number> = {};
-
-      failures.forEach(failure => {
-        const parts = failure.testName.split('_');
-        // Format: TestName_category_parts_index
-        // First part is testName (may contain dots), last part is index
-        // Everything in between is the category (may have underscores like bug_critical)
-        if (parts.length >= 3) {
-          const categoryParts = parts.slice(1, -1); // Remove first (testName) and last (index)
-          const category = categoryParts.join('_');
-          categoryCounts[category] = (categoryCounts[category] || 0) + 1;
-        } else {
-          categoryCounts['unknown'] = (categoryCounts['unknown'] || 0) + 1;
-        }
-      });
-
-      const total = failures.length;
-
-      const categoryColors: Record<FailureCategory, string> = {
-        bug_critical: '#ef4444',
-        bug_minor: '#f59e0b',
-        environment: '#3b82f6',
-        flaky: '#8b5cf6',
-        configuration: '#10b981',
-        unknown: '#64748b',
-      };
-
-      // Convert to array and sort by count
-      return Object.entries(categoryCounts).map(([category, count]) => {
-        return {
-          category: category as FailureCategory,
-          count,
-          percentage: total > 0 ? (count / total) * 100 : 0,
-          color: categoryColors[category as FailureCategory] || '#64748b',
-        };
-      }).sort((a, b) => b.count - a.count);
-    } catch (error) {
-      console.error('Failed to get failure categories breakdown:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Helper: Get similar failures count
-   */
-  private static async getSimilarFailuresCount(): Promise<number> {
-    try {
-      // Would query vector DB for similar failures
-      // Mock for now
-      return 89;
-    } catch (error) {
-      console.error('Failed to get similar failures count:', error);
-      return 0;
-    }
-  }
-
-  /**
-   * Helper: Get provider comparison data
-   */
-  private static getProviderComparison(): DashboardMetrics['providers'] {
-    const activeProvider = process.env.AI_PROVIDER || 'anthropic';
-
-    return [
-      {
-        name: 'Claude Opus 4.6',
-        costPer1M: 15.00,
-        contextWindow: '200K',
-        speed: '⚡⚡⚡',
-        isActive: activeProvider === 'anthropic',
-      },
-      {
-        name: 'GPT-4.1',
-        costPer1M: 8.00,
-        contextWindow: '1M',
-        speed: '⚡⚡⚡',
-        isActive: activeProvider === 'openai',
-      },
-      {
-        name: 'Gemini 3.0 Flash',
-        costPer1M: 0.15,
-        contextWindow: '1M',
-        speed: '⚡⚡⚡⚡',
-        isActive: activeProvider === 'google',
-      },
-      {
-        name: 'OpenRouter (200+ models)',
-        costPer1M: 0.00,
-        contextWindow: 'Varies',
-        speed: '⚡⚡⚡',
-        isActive: activeProvider === 'openrouter',
-      },
-    ];
-  }
-
-  /**
-   * Helper: Get start date from time range
-   */
-  private static getStartDate(timeRange: string): Date {
-    const now = new Date();
-
-    switch (timeRange) {
-      case '1h':
-        return new Date(now.getTime() - 60 * 60 * 1000);
-      case '24h':
-        return new Date(now.getTime() - 24 * 60 * 60 * 1000);
-      case '7d':
-        return new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
-      case '30d':
-        return new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000);
-      default:
-        return new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000);
-    }
-  }
-
-  /**
-   * Helper: Get time range label
-   */
-  private static getTimeRangeLabel(timeRange: string): string {
-    switch (timeRange) {
-      case '1h':
-        return 'Last hour';
-      case '24h':
-        return 'Last 24 hours';
-      case '7d':
-        return 'Last 7 days';
-      case '30d':
-        return 'Last 30 days';
-      default:
-        return 'Last 30 days';
     }
   }
 }
