+++ backend/src/controllers/notification.controller.ts
@@ -1,245 +1,54 @@
-import { prisma } from '@/lib/prisma';
-import { NotFoundError } from '@/middleware/errorHandler';
-import { NotificationService } from '@/services/notification.service';
-import { logger } from '@/utils/logger';
+/**
+ * NotificationController — Thin HTTP adapter.
+ *
+ * Delegates all notification preferences, history, and delivery to
+ * NotificationPreferenceService. No Prisma imports allowed here.
+ */
 
-// Types for user notification preferences (stored as JSON)
-interface NotificationPreferences {
-  email?: {
-    enabled: boolean;
-    address: string;
-    digest: boolean;
-    digestFrequency: 'daily' | 'weekly';
-  };
-  slack?: {
-    enabled: boolean;
-    channel: string;
-    mentions: string[];
-  };
-  pushover?: {
-    enabled: boolean;
-    deviceGroups: string[];
-    priority: number;
-  };
-  conditions: {
-    pipelineStart: boolean;
-    pipelineSuccess: boolean;
-    pipelineFailure: boolean;
-    testFlaky: boolean;
-    coverageDecrease: boolean;
-  };
-}
-
-interface NotificationFilters {
-  startDate?: string;
-  endDate?: string;
-  channel?: string;
-  status?: string;
-  type?: string;
-}
+import {
+  notificationPreferenceService,
+  NotificationPreferences,
+  NotificationFilters,
+} from '../services/notification-preference.service';
 
 export class NotificationController {
-  private notificationService: NotificationService;
-
-  constructor() {
-    this.notificationService = new NotificationService();
+  async getPreferences(userId: string) {
+    return notificationPreferenceService.getPreferences(userId);
   }
 
-  async getPreferences(_userId: string): Promise<NotificationPreferences> {
-    // We don't have a NotificationPreference model in Prisma schema I added?
-    // I added: Notification, TestRailRun, ConfluencePage.
-    // I did NOT add NotificationPreference.
-    // The previous controller used `NotificationPreference` model.
-    // It seems I missed this model in my schema update.
-    // Strategy: Use User model's `metadata` or imply default preferences for now.
-    // Or just better: Return default preferences to avoid breakage if unrelated to fixing TestRun.
-    // Re-reading error: `NotificationController: missing models`.
-    // I should strictly "fix" it.
-    // I will return default preferences logic as fallback.
-
-    return {
-      conditions: {
-        pipelineStart: false,
-        pipelineSuccess: false,
-        pipelineFailure: true,
-        testFlaky: true,
-        coverageDecrease: true,
-      },
-    };
+  async updatePreferences(userId: string, preferences: NotificationPreferences) {
+    return notificationPreferenceService.updatePreferences(userId, preferences);
   }
 
-  async updatePreferences(userId: string, preferences: NotificationPreferences): Promise<NotificationPreferences> {
-    // Stub implementation as model is missing
-    logger.info(`Updated notification preferences for user: ${userId}`);
-    return preferences;
+  async sendTestNotification(userId: string) {
+    return notificationPreferenceService.sendTestNotification(userId);
   }
 
-  async sendTestNotification(userId: string): Promise<void> {
-    const user = await prisma.user.findUnique({ where: { id: userId } });
-    if (!user) {
-      throw new NotFoundError('User not found');
-    }
-
-    const testMessage = 'This is a test notification from TestOps Companion';
-
-    // Use service to send
-    await this.notificationService.sendNotifications({
-      enabled: true,
-      channels: ['email'], // Default channel for test
-      message: testMessage,
-      userId,
-    } as { enabled: boolean; channels: string[]; message?: string; userId?: string }, testMessage);
-
-    // Record history
-    await prisma.notification.create({
-      data: {
-        userId,
-        type: 'SYSTEM', // String literal for Dev schema compatibility
-        title: 'Test Notification',
-        message: testMessage,
-
-      }
-    });
-
-    logger.info(`Test notification sent to user: ${userId}`);
+  async getAvailableChannels() {
+    return notificationPreferenceService.getAvailableChannels();
   }
 
-  async getAvailableChannels(): Promise<Array<{ id: string; name: string; description: string; configSchema: Record<string, unknown>; enabled: boolean }>> {
-    return [
-      {
-        id: 'email',
-        name: 'Email',
-        description: 'Send notifications via email',
-        configSchema: {
-          address: 'string',
-          digest: 'boolean',
-          digestFrequency: ['daily', 'weekly'],
-        },
-        enabled: true,
-      },
-      {
-        id: 'slack',
-        name: 'Slack',
-        description: 'Send notifications to Slack channels',
-        configSchema: {
-          channel: 'string',
-          mentions: 'array',
-        },
-        enabled: true,
-      },
-      {
-        id: 'pushover',
-        name: 'Pushover',
-        description: 'Send notifications via Pushover',
-        configSchema: {
-          deviceGroups: 'array',
-          priority: 'number',
-        },
-        enabled: true,
-      },
-    ];
+  async verifyChannel(userId: string, channelConfig: unknown) {
+    return notificationPreferenceService.verifyChannel(userId, channelConfig);
   }
 
-  async verifyChannel(_userId: string, channelConfig: unknown): Promise<{
-    valid: boolean;
-    message?: string;
-  }> {
-    try {
-      await this.notificationService.verifyChannelConfig(channelConfig);
-      return { valid: true };
-    } catch (error) {
-      return {
-        valid: false,
-        message: (error as Error).message,
-      };
-    }
+  async getNotificationHistory(userId: string, filters: NotificationFilters) {
+    return notificationPreferenceService.getHistory(userId, filters);
   }
 
-  async getNotificationHistory(userId: string, filters: NotificationFilters): Promise<unknown[]> {
-    const where: Record<string, unknown> = { userId };
-
-    // Basic mapping of filters to Prisma
-    if (filters.status) {
-      // Our Notification model doesn't have explicit status enum, maybe inside 'data' JSON
-      // or we added 'read' boolean.
-      // The previous Sequelize model had 'status' string.
-      // My added Prisma model has: type, title, message, read, data.
-      // Maybe I should filter by 'read' if status is related?
-      // Leaving as is: strict filter likely fails if field missing.
-    }
-
-    if (filters.type) {
-      where.type = filters.type;
-    }
-
-    if (filters.startDate || filters.endDate) {
-      const createdAt: { gte?: Date; lte?: Date } = {};
-      if (filters.startDate) {
-        createdAt.gte = new Date(filters.startDate);
-      }
-      if (filters.endDate) {
-        createdAt.lte = new Date(filters.endDate);
-      }
-      where.createdAt = createdAt;
-    }
-
-    return prisma.notification.findMany({
-      where,
-      orderBy: { createdAt: 'desc' },
-    });
+  async getDeliveryMetrics() {
+    return notificationPreferenceService.getDeliveryMetrics();
   }
 
-  async getDeliveryMetrics(): Promise<Record<string, unknown>> {
-    const [
-      totalNotifications,
-      // Status field missing in my new model, removing status based counts
-    ] = await Promise.all([
-      prisma.notification.count(),
-    ]);
-
-    return {
-      totalNotifications,
-      successfulDeliveries: totalNotifications, // Mocking as we lost status tracking details
-      failedDeliveries: 0,
-      deliveryRate: 100,
-      channelStats: {},
-    };
+  async sendBroadcastNotification(data: { message: string; channels: string[]; userGroups?: string[] }) {
+    return notificationPreferenceService.sendBroadcast(data);
   }
 
-  async sendBroadcastNotification(data: {
-    message: string;
-    channels: string[];
-    userGroups?: string[];
-  }): Promise<void> {
-    const where = data.userGroups ? { role: { in: data.userGroups } } as Record<string, unknown> : {};
-    const users = await prisma.user.findMany({ where });
-
-    for (const user of users) {
-      await this.notificationService.sendNotifications({
-        enabled: true,
-        channels: data.channels as Array<'slack' | 'email' | 'pushover'>,
-        message: data.message,
-        userId: user.id,
-      } as { enabled: boolean; channels: string[]; message?: string; userId?: string }, data.message);
-
-      await prisma.notification.create({
-        data: {
-          userId: user.id,
-          type: 'SYSTEM', // String literal for Dev schema compatibility
-          title: 'Broadcast',
-          message: data.message,
-        }
-      });
-    }
-
-    logger.info(`Broadcast notification sent to ${users.length} users`);
+  async getGlobalSettings() {
+    return notificationPreferenceService.getGlobalSettings();
   }
 
-  async getGlobalSettings(): Promise<Record<string, unknown>> {
-    return {};
-  }
-
-  async updateGlobalSettings(settings: Record<string, unknown>): Promise<Record<string, unknown>> {
-    return settings;
+  async updateGlobalSettings(settings: Record<string, unknown>) {
+    return notificationPreferenceService.updateGlobalSettings(settings);
   }
 }
