+++ backend/src/controllers/pipeline.controller.ts
@@ -1,264 +1,60 @@
-import { prisma } from '../lib/prisma';
-import { auditService } from '../services/audit.service';
-import { GitHubService } from '../services/github.service';
-import { NotFoundError } from '../middleware/errorHandler';
+/**
+ * PipelineController — Thin HTTP adapter.
+ *
+ * Delegates all business logic and data access to PipelineService.
+ * No Prisma imports allowed here.
+ */
+
+import { pipelineService } from '../services/pipeline.service';
 import {
   CreatePipelineDTO,
   UpdatePipelineDTO,
   PipelineFilters,
-  toPipeline
 } from '../types/pipeline';
-import { TestStatus, PipelineType } from '../constants';
-import {
-  createPipelineInput,
-  updatePipelineInput,
-  parsePipelineConfig
-} from '../utils/prismaHelpers';
 
 export class PipelineController {
-  private githubService: GitHubService;
-
-  constructor() {
-    this.githubService = new GitHubService();
-  }
-
   async listPipelines(userId: string, filters?: PipelineFilters) {
-    const where = {
-      // userId, // Removed
-      ...(filters?.type && { type: filters.type }),
-      // ...(filters?.status && { status: filters.status }) // Removed status
-    };
-
-    const prismaPipelines = await prisma.pipeline.findMany({
-      where,
-      include: {
-        testRuns: {
-          orderBy: { createdAt: 'desc' },
-          take: 20
-        }
-      }
-    });
-
-    // Transform to frontend format with calculated success rate
-    return prismaPipelines.map(pipeline => {
-      const testRuns = pipeline.testRuns || [];
-      const lastRun = testRuns.length > 0 ? testRuns[0].createdAt.toISOString() : new Date().toISOString();
-
-      // Calculate success rate
-      const totalRuns = testRuns.length;
-      const passedRuns = testRuns.filter(run => run.status === 'PASSED').length;
-      const successRate = totalRuns > 0 ? Math.round((passedRuns / totalRuns) * 100) : 0;
-
-      // Map status
-      const lastRunStatus = testRuns.length > 0 ? testRuns[0].status : 'PENDING';
-      const statusMap: Record<string, string> = {
-        'PASSED': 'success',
-        'FAILED': 'failed',
-        'RUNNING': 'running',
-        'PENDING': 'pending'
-      };
-
-      const rawConfig = typeof pipeline.config === 'string' ? JSON.parse(pipeline.config) : pipeline.config;
-      // Strip credentials from API response
-      const { credentials: _creds, ...safeConfig } = rawConfig || {};
-
-      return {
-        id: pipeline.id,
-        name: pipeline.name,
-        type: pipeline.type.toLowerCase().replace('_', '-') as 'jenkins' | 'github-actions',
-        status: statusMap[lastRunStatus] || 'pending',
-        lastRun,
-        successRate,
-        config: safeConfig
-      };
-    });
+    return pipelineService.list(userId, filters);
   }
 
-  async getPipeline(id: string, _userId: string) {
-    const prismaPipeline = await prisma.pipeline.findUnique({
-      where: { id },
-      include: {
-        testRuns: {
-          orderBy: { createdAt: 'desc' },
-          take: 5
-        }
-      }
-    });
-
-    if (!prismaPipeline) {
-      throw new NotFoundError('Pipeline not found');
-    }
-
-    // if (prismaPipeline.userId !== userId) {
-    //   throw new AuthorizationError('Not authorized to access this pipeline');
-    // }
-
-    // Convert Prisma pipeline to our Pipeline type
-    return toPipeline(prismaPipeline);
+  async getPipeline(id: string, userId: string) {
+    return pipelineService.getById(id, userId);
   }
 
   async createPipeline(data: CreatePipelineDTO, userId: string) {
-    // Validate connection before creating
-    await this.githubService.validateConnection(data.config);
-
-    const createData = createPipelineInput({
-      ...data,
-      // userId, // Removed
-      // status: PipelineStatus.PENDING // Removed
-    });
-
-    const prismaPipeline = await prisma.pipeline.create({
-      data: createData
-    });
-
-    // Audit Log
-    void auditService.log(
-      'PIPELINE_CREATE',
-      'Pipeline',
-      prismaPipeline.id,
-      userId,
-      { name: prismaPipeline.name, type: prismaPipeline.type }
-    );
-
-    // Convert Prisma pipeline to our Pipeline type
-    return toPipeline(prismaPipeline);
+    return pipelineService.create(data, userId);
   }
 
   async updatePipeline(id: string, data: UpdatePipelineDTO, userId: string) {
-    await this.getPipeline(id, userId);
-
-    if (data.config) {
-      await this.githubService.validateConnection(data.config);
-    }
-
-    const updateData = updatePipelineInput(data);
-
-    const prismaPipeline = await prisma.pipeline.update({
-      where: { id },
-      data: updateData
-    });
-
-    // Audit Log
-    void auditService.log(
-      'PIPELINE_UPDATE',
-      'Pipeline',
-      prismaPipeline.id,
-      userId,
-      { updates: Object.keys(data) }
-    );
-
-    // Convert Prisma pipeline to our Pipeline type
-    return toPipeline(prismaPipeline);
+    return pipelineService.update(id, data, userId);
   }
 
   async deletePipeline(id: string, userId: string) {
-    await this.getPipeline(id, userId);
-
-    const deleted = await prisma.pipeline.delete({
-      where: { id }
-    });
-
-    // Audit Log
-    void auditService.log(
-      'PIPELINE_DELETE',
-      'Pipeline',
-      id,
-      userId,
-      { name: deleted.name }
-    );
+    return pipelineService.delete(id, userId);
   }
 
   async startPipeline(id: string, userId: string) {
-    const pipeline = await this.getPipeline(id, userId);
-
-    // Start pipeline based on type
-    let testRun;
-
-    switch (pipeline.type) {
-      case PipelineType.GITHUB_ACTIONS:
-        testRun = await this.githubService.startPipeline(pipeline);
-        break;
-      default:
-        throw new Error(`Unsupported pipeline type: ${pipeline.type}`);
-    }
-
-    // Update pipeline status - REMOVED as status field is gone
-    /* await prisma.pipeline.update({
-      where: { id },
-      data: { status: 'RUNNING' }
-    }); */
-
-    return testRun;
+    return pipelineService.start(id, userId);
   }
 
   async getTestRuns(id: string, userId: string) {
-    const pipeline = await this.getPipeline(id, userId);
-
-    const testRuns = await prisma.testRun.findMany({
-      where: { pipelineId: pipeline.id },
-      orderBy: { createdAt: 'desc' },
-      take: 100
-    });
-
-    return testRuns;
+    return pipelineService.getTestRuns(id, userId);
   }
 
   async schedulePipeline(id: string, schedule: string, userId: string) {
-    const pipeline = await this.getPipeline(id, userId);
-    const currentConfig = parsePipelineConfig<Record<string, unknown>>(pipeline.config);
-
-    await prisma.pipeline.update({
-      where: { id },
-      data: {
-        config: JSON.stringify({
-          ...currentConfig,
-          schedule
-        })
-      }
-    });
+    return pipelineService.schedule(id, schedule, userId);
   }
 
   async getFailedTests(id: string, userId: string) {
-    const pipeline = await this.getPipeline(id, userId);
-
-    const testRuns = await prisma.testRun.findMany({
-      where: {
-        pipelineId: pipeline.id,
-        status: TestStatus.FAILED
-      },
-      orderBy: { createdAt: 'desc' }
-    });
-
-    return testRuns;
+    return pipelineService.getFailedTests(id, userId);
   }
 
   async getFlakeyTests(id: string, userId: string) {
-    const pipeline = await this.getPipeline(id, userId);
-
-    const testRuns = await prisma.testRun.findMany({
-      where: {
-        pipelineId: pipeline.id,
-        OR: [
-          { status: TestStatus.FAILED },
-          { status: TestStatus.FLAKY }
-        ]
-      },
-      orderBy: { createdAt: 'desc' }
-    });
-
-    return testRuns;
+    return pipelineService.getFlakyTests(id, userId);
   }
 
   async validatePipelineConfig(config: unknown) {
-    try {
-      await this.githubService.validateConnection(config);
-      return { valid: true };
-    } catch (error) {
-      return {
-        valid: false,
-        error: error instanceof Error ? error.message : 'Invalid configuration'
-      };
-    }
+    return pipelineService.validateConfig(config);
   }
 }
