+++ backend/src/routes/pipeline.routes.ts
@@ -3,7 +3,6 @@ import { asyncHandler } from '../middleware/errorHandler';
 import { pipelineController } from '../controllers/pipeline.controller';
 import { UserRole, PipelineType } from '../constants';
 import { pipelineRouter as router } from './index';
-import { prisma } from '../lib/prisma';
 
 // Apply authentication to all routes
 router.use(authenticate);
@@ -124,34 +123,8 @@ router.get(
       return;
     }
 
-    const runs = await prisma.testRun.findMany({
-      where: { pipelineId: req.params.id as string, userId: req.user.id },
-      orderBy: { createdAt: 'desc' },
-      take: 20,
-      include: { results: true }
-    });
-
-    const statusMap: Record<string, string> = {
-      'PASSED': 'success',
-      'FAILED': 'failed',
-      'RUNNING': 'running',
-      'PENDING': 'pending'
-    };
-
-    const formattedRuns = runs.map(run => {
-      // Count failed results
-      const failed = run.results?.filter(r => r.status === 'FAILED').length || 0;
-      return {
-        id: run.id,
-        status: statusMap[run.status] || 'pending',
-        startTime: run.startedAt?.toISOString() || run.createdAt.toISOString(),
-        endTime: run.completedAt?.toISOString() || run.createdAt.toISOString(),
-        duration: run.duration || 0,
-        errorCount: failed
-      };
-    });
-
-    res.json(formattedRuns);
+    const runs = await pipelineController.getTestRuns(req.params.id as string, req.user.id);
+    res.json(runs);
   })
 );
