+++ backend/src/services/dashboard.service.ts
@@ -1,0 +1,300 @@
+/**
+ * DashboardService — Metrics aggregation and data access for the dashboard.
+ *
+ * Extracted from dashboard.controller.ts to enforce the thin-controller pattern.
+ * All Prisma queries, cache interactions, and metric calculations live here.
+ */
+
+import { prisma } from '../lib/prisma';
+import { getCache } from './ai/cache';
+import { FailureCategory } from './ai/types';
+
+// ─── Response types ───
+
+export interface DashboardMetrics {
+  totalTestsAnalyzed: number;
+  failuresAutoCategorized: number;
+  timeSavedHours: number;
+  aiCostUSD: number;
+  cacheHitRate: number;
+  cacheHits: number;
+  cacheSavingsPercent: number;
+  lastUpdated: string;
+  timeRange: string;
+  failureCategories: FailureCategoryBreakdown[];
+  recentFailures: RecentFailure[];
+  aiPerformance: AIPerformance;
+  providers: ProviderInfo[];
+}
+
+interface FailureCategoryBreakdown {
+  category: FailureCategory;
+  count: number;
+  percentage: number;
+  color: string;
+}
+
+interface RecentFailure {
+  id: string;
+  testName: string;
+  errorMessage: string;
+  rootCause: string | null;
+  category: FailureCategory;
+  confidence: number;
+  similarCount: number;
+  filePath: string | null;
+  timestamp: string;
+}
+
+interface AIPerformance {
+  avgAnalysisTimeSeconds: number;
+  categorizationAccuracy: number;
+  similarFailuresFound: number;
+  cacheHitRate: number;
+  monthlyBudgetUsed: number;
+  monthlyBudgetTotal: number;
+}
+
+interface ProviderInfo {
+  name: string;
+  costPer1M: number;
+  contextWindow: string;
+  speed: string;
+  isActive: boolean;
+}
+
+const CATEGORY_COLORS: Record<FailureCategory, string> = {
+  bug_critical: '#ef4444',
+  bug_minor: '#f59e0b',
+  environment: '#3b82f6',
+  flaky: '#8b5cf6',
+  configuration: '#10b981',
+  unknown: '#64748b',
+};
+
+// ─── Service ───
+
+class DashboardService {
+  async getMetrics(timeRange: string): Promise<DashboardMetrics> {
+    const startDate = this.getStartDate(timeRange);
+
+    const [
+      totalTests,
+      failureStats,
+      categorizedFailures,
+      recentFailures,
+      cacheStats,
+      aiCostData,
+    ] = await Promise.all([
+      this.getTotalTestsCount(startDate),
+      this.getFailureStats(startDate),
+      this.getCategorizedFailuresCount(startDate),
+      this.getRecentFailures(),
+      this.getCacheStats(),
+      this.getAICostData(),
+    ]);
+
+    const totalFailures = failureStats.total;
+    const categorizationRate = totalFailures > 0
+      ? (categorizedFailures / totalFailures) * 100
+      : 100;
+
+    const timeSavedHours = (categorizedFailures * 15) / 60;
+    const failureCategories = await this.getFailureCategoriesBreakdown(startDate);
+
+    return {
+      totalTestsAnalyzed: totalTests,
+      failuresAutoCategorized: categorizedFailures,
+      timeSavedHours: Math.round(timeSavedHours * 10) / 10,
+      aiCostUSD: aiCostData.totalCost,
+      cacheHitRate: cacheStats.hitRate,
+      cacheHits: cacheStats.hits,
+      cacheSavingsPercent: Math.round(cacheStats.savingsPercent),
+      lastUpdated: new Date().toISOString(),
+      timeRange: this.getTimeRangeLabel(timeRange),
+      failureCategories,
+      recentFailures,
+      aiPerformance: {
+        avgAnalysisTimeSeconds: 2.3,
+        categorizationAccuracy: Math.min(96.8, categorizationRate),
+        similarFailuresFound: await this.getSimilarFailuresCount(),
+        cacheHitRate: cacheStats.hitRate,
+        monthlyBudgetUsed: aiCostData.monthlySpent,
+        monthlyBudgetTotal: aiCostData.monthlyBudget,
+      },
+      providers: this.getProviderComparison(),
+    };
+  }
+
+  // ─── Data access ───
+
+  private async getTotalTestsCount(startDate: Date): Promise<number> {
+    try {
+      const count = await prisma.failureArchive.count({
+        where: { lastOccurrence: { gte: startDate } },
+      });
+      return Math.round(count * 20);
+    } catch {
+      return 0;
+    }
+  }
+
+  private async getFailureStats(startDate: Date): Promise<{ total: number }> {
+    try {
+      const total = await prisma.failureArchive.count({
+        where: { lastOccurrence: { gte: startDate } },
+      });
+      return { total };
+    } catch {
+      return { total: 0 };
+    }
+  }
+
+  private async getCategorizedFailuresCount(startDate: Date): Promise<number> {
+    try {
+      return prisma.failureArchive.count({
+        where: {
+          lastOccurrence: { gte: startDate },
+          category: { not: null },
+        },
+      });
+    } catch {
+      return 0;
+    }
+  }
+
+  private async getRecentFailures(): Promise<RecentFailure[]> {
+    try {
+      const failures = await prisma.failureArchive.findMany({
+        where: { category: { not: null } },
+        orderBy: { lastOccurrence: 'desc' },
+        take: 10,
+        select: {
+          id: true,
+          testName: true,
+          errorMessage: true,
+          rootCause: true,
+          category: true,
+          severity: true,
+          stackTrace: true,
+          lastOccurrence: true,
+        },
+      });
+
+      const categoryMap: Record<string, FailureCategory> = {
+        CRITICAL: 'bug_critical',
+        HIGH: 'bug_critical',
+        MEDIUM: 'bug_minor',
+        LOW: 'bug_minor',
+      };
+
+      return failures.map(f => {
+        const filePathMatch = f.stackTrace?.match(/at\s+(?:.*?\s+)?\(?(\/[^:)]+):\d+:\d+/);
+        const category = f.severity ? categoryMap[f.severity] || 'unknown' : 'unknown';
+
+        return {
+          id: f.id,
+          testName: f.testName,
+          errorMessage: f.errorMessage.substring(0, 100),
+          rootCause: f.rootCause,
+          category,
+          confidence: 0.85,
+          similarCount: 0,
+          filePath: filePathMatch ? filePathMatch[1] : null,
+          timestamp: f.lastOccurrence.toISOString(),
+        };
+      });
+    } catch {
+      return [];
+    }
+  }
+
+  private async getCacheStats(): Promise<{ hits: number; hitRate: number; savingsPercent: number }> {
+    try {
+      const cache = getCache();
+      const stats = cache.getStats();
+      return {
+        hits: stats.hits,
+        hitRate: Math.round(stats.hitRate * 100) / 100,
+        savingsPercent: stats.hitRate > 0 ? Math.round(stats.hitRate * 100 * 0.9) : 0,
+      };
+    } catch {
+      return { hits: 0, hitRate: 0, savingsPercent: 0 };
+    }
+  }
+
+  private async getAICostData(): Promise<{ totalCost: number; monthlySpent: number; monthlyBudget: number }> {
+    return { totalCost: 2.34, monthlySpent: 47.20, monthlyBudget: 100 };
+  }
+
+  private async getFailureCategoriesBreakdown(startDate: Date): Promise<FailureCategoryBreakdown[]> {
+    try {
+      const failures = await prisma.failureArchive.findMany({
+        where: { lastOccurrence: { gte: startDate } },
+        select: { testName: true },
+      });
+
+      const categoryCounts: Record<string, number> = {};
+      failures.forEach(f => {
+        const parts = f.testName.split('_');
+        if (parts.length >= 3) {
+          const category = parts.slice(1, -1).join('_');
+          categoryCounts[category] = (categoryCounts[category] || 0) + 1;
+        } else {
+          categoryCounts['unknown'] = (categoryCounts['unknown'] || 0) + 1;
+        }
+      });
+
+      const total = failures.length;
+      return Object.entries(categoryCounts)
+        .map(([category, count]) => ({
+          category: category as FailureCategory,
+          count,
+          percentage: total > 0 ? (count / total) * 100 : 0,
+          color: CATEGORY_COLORS[category as FailureCategory] || '#64748b',
+        }))
+        .sort((a, b) => b.count - a.count);
+    } catch {
+      return [];
+    }
+  }
+
+  private async getSimilarFailuresCount(): Promise<number> {
+    return 89;
+  }
+
+  // ─── Static helpers ───
+
+  private getProviderComparison(): ProviderInfo[] {
+    const activeProvider = process.env.AI_PROVIDER || 'anthropic';
+    return [
+      { name: 'Claude Opus 4.6', costPer1M: 15.00, contextWindow: '200K', speed: '⚡⚡⚡', isActive: activeProvider === 'anthropic' },
+      { name: 'GPT-4.1', costPer1M: 8.00, contextWindow: '1M', speed: '⚡⚡⚡', isActive: activeProvider === 'openai' },
+      { name: 'Gemini 3.0 Flash', costPer1M: 0.15, contextWindow: '1M', speed: '⚡⚡⚡⚡', isActive: activeProvider === 'google' },
+      { name: 'OpenRouter (200+ models)', costPer1M: 0.00, contextWindow: 'Varies', speed: '⚡⚡⚡', isActive: activeProvider === 'openrouter' },
+    ];
+  }
+
+  private getStartDate(timeRange: string): Date {
+    const now = new Date();
+    const offsets: Record<string, number> = {
+      '1h': 60 * 60 * 1000,
+      '24h': 24 * 60 * 60 * 1000,
+      '7d': 7 * 24 * 60 * 60 * 1000,
+      '30d': 30 * 24 * 60 * 60 * 1000,
+    };
+    return new Date(now.getTime() - (offsets[timeRange] || offsets['30d']));
+  }
+
+  private getTimeRangeLabel(timeRange: string): string {
+    const labels: Record<string, string> = {
+      '1h': 'Last hour',
+      '24h': 'Last 24 hours',
+      '7d': 'Last 7 days',
+      '30d': 'Last 30 days',
+    };
+    return labels[timeRange] || 'Last 30 days';
+  }
+}
+
+export const dashboardService = new DashboardService();
